requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
mysql-connector-python==8.2.0
pyjwt==2.8.0
requests==2.31.0
python-dotenv==1.0.0
email-validator==2.1.0
asyncio-mqtt==0.16.1
twilio==8.10.0
python-telegram-bot==20.7
python-dateutil==2.8.2
pytz==2023.3
httpx==0.25.2

# Development/test dependencies are intentionally omitted from this file
# to keep the production image slim. Add them to a separate
# requirements-dev.txt if needed for CI/testing.
